+++ routes/withdrawRouter.js
@@ -1,15 +1,17 @@
-/* eslint-disable no-throw-literal */
 const express = require('express')
 const router = express.Router()
-const withdrawService = require('../services/withdrawService')
+const withdrawService = require('../services/withdraw')
 
 const availableNotes = [100, 50, 20, 10]
 
+// TODO userInput name
+
 router.get('/', (req, res, next) => {
+  console.log(req.query.amount)
   res.render('withdraw')
 })
 
 router.post('/', (req, res, next) => {
-  withdrawService(req.body.amount, availableNotes, res)
+  withdrawService.withdraw(req.body.amount, availableNotes, res)
 })
 module.exports = router
